add result api resource

+++ src/main/kotlin/com/example/shared/schema/models/Location.kt
@@ -3,8 +3,8 @@ package com.example.shared.schema.models
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class Location(
-    @JsonProperty("lat") val latitude: Double,
-    @JsonProperty("long") val longitude: Double,
+    @JsonProperty("lat") val latitude: Float,
+    @JsonProperty("long") val longitude: Float,
     val locality: String,
     val country: String,
 )
